check for customer portal link
