+++ .env.client.example
@@ -1,3 +1,2 @@
 # learn more about client side env vars https://wasp-lang.dev/docs/project/env-vars
-
 REACT_APP_SOME_VAR_NAME=foo
